city spinner functional and weather info show properly

+++ app/src/main/java/com/hellohasan/weatherforecast/features/weather_info_show/model/data_class/Main.kt
@@ -7,7 +7,7 @@ data class Main(
         @SerializedName("temp")
         val temp: Double = 0.0,
         @SerializedName("pressure")
-        val pressure: Int = 0,
+        val pressure: Double = 0.0,
         @SerializedName("humidity")
         val humidity: Int = 0,
         @SerializedName("temp_min")
